Fix some debug information

+++ src/xpn_client/nfi/nfi_server.hpp
@@ -109,7 +109,7 @@ namespace XPN
                 return -1;
             }
 
-            debug_info("[NFI_XPN] [nfi_write_operation] Execute operation: "<<op<<" -> ");
+            debug_info("[NFI_XPN] [nfi_write_operation] Execute operation: "<<static_cast<int>(op)<<" -> ");
             ret = m_comm->write_data((void *)&(msg), sizeof(msg));
 
             debug_info("[NFI_XPN] [nfi_write_operation] >> End");
@@ -128,7 +128,7 @@ namespace XPN
             }
 
             // send request...
-            debug_info("[NFI_XPN] [nfi_server_do_request] Send operation: "<<op);
+            debug_info("[NFI_XPN] [nfi_server_do_request] Send operation: "<<static_cast<int>(op));
 
             ret = nfi_write_operation(op, msg);
             if (ret < 0) {
@@ -136,7 +136,7 @@ namespace XPN
             }
 
             // read response...
-            debug_info("[NFI_XPN] [nfi_server_do_request] Response operation: "<<op);
+            debug_info("[NFI_XPN] [nfi_server_do_request] Response operation: "<<static_cast<int>(op));
 
             ret = m_comm->read_data((void *)&(req), sizeof(req));
             if (ret < 0) {
